requirements.txt
Django==2.1.2
django-bulk-update==2.2.0
django-celery-beat==1.5.0
django-cors-middleware==1.3.1
django-oauth-toolkit==1.2.0
django-pyodbc-azure==2.1.0.0
django-rest-passwordreset==0.9.5
django-seed==0.1.9
django-timezone-field==3.0
django-wkhtmltopdf==3.2.0
djangorestframework==3.8.2
pika
autobahn
twisted
PyYAML==3.13
mysql-connector==2.1.7
mysqlclient==1.3.13
requests==2.18.4
validators